this took way too long

+++ code/game/jobs/job_controller.dm
@@ -371,8 +371,7 @@ var/global/datum/controller/occupations/job_master
 			if(istype(S, /obj/effect/landmark/start) && istype(S.loc, /turf))
 				H.forceMove(S.loc)
 			else
-				var/list/spawn_props = LateSpawn(H.client, rank)
-				var/turf/T = spawn_props["turf"]
+				var/turf/T = get_turf(pick(latejoin))
 				if(!T)
 					to_chat(H, "<span class='critical'>You were unable to be spawned at your chosen late-join spawnpoint. Please verify your job/spawn point combination makes sense, and try another one.</span>")
 					return
@@ -511,21 +510,22 @@ var/global/datum/controller/occupations/job_master
 					to_chat(H, "<span class='danger'>Failed to locate a storage object on your mob, either you spawned with no arms and no backpack or this is a bug.</span>")
 
 		if(istype(H)) //give humans wheelchairs, if they need them.
-			var/obj/item/organ/external/l_foot = H.get_organ("l_foot")
-			var/obj/item/organ/external/r_foot = H.get_organ("r_foot")
-			var/obj/item/weapon/storage/S = locate() in H.contents
-			var/obj/item/wheelchair/R = null
-			if(S)
-				R = locate() in S.contents
-			if(!l_foot || !r_foot || R)
-				var/obj/structure/bed/chair/wheelchair/W = new /obj/structure/bed/chair/wheelchair(H.loc)
-				W.buckle_mob(H)
-				H.update_canmove()
-				W.set_dir(H.dir)
-				W.add_fingerprint(H)
-				if(R)
-					W.color = R.color
-					qdel(R)
+			if(istype(H.loc, /obj/belly)) //unless in a gut
+				var/obj/item/organ/external/l_foot = H.get_organ("l_foot")
+				var/obj/item/organ/external/r_foot = H.get_organ("r_foot")
+				var/obj/item/weapon/storage/S = locate() in H.contents
+				var/obj/item/wheelchair/R = null
+				if(S)
+					R = locate() in S.contents
+				if(!l_foot || !r_foot || R)
+					var/obj/structure/bed/chair/wheelchair/W = new /obj/structure/bed/chair/wheelchair(H.loc)
+					W.buckle_mob(H)
+					H.update_canmove()
+					W.set_dir(H.dir)
+					W.add_fingerprint(H)
+					if(R)
+						W.color = R.color
+						qdel(R)
 
 		to_chat(H, "<B>You are [job.total_positions == 1 ? "the" : "a"] [alt_title ? alt_title : rank].</B>")
 
@@ -649,24 +649,74 @@ var/global/datum/controller/occupations/job_master
 
 	var/datum/spawnpoint/spawnpos
 	var/fail_deadly = FALSE
+	var/obj/belly/vore_spawn_gut
 
 	var/datum/job/J = SSjob.get_job(rank)
 	fail_deadly = J?.offmap_spawn
 
 	//Spawn them at their preferred one
 	if(C && C.prefs.spawnpoint)
-		if(!(C.prefs.spawnpoint in using_map.allowed_spawns))
-			if(fail_deadly)
-				to_chat(C, "<span class='warning'>Your chosen spawnpoint is unavailable for this map and your job requires a specific spawnpoint. Please correct your spawn point choice.</span>")
-				return
+		if(C.prefs.spawnpoint == "Vore Belly")
+			var/list/preds = list()
+			var/list/pred_names = list() //This is cringe
+			for(var/client/V in GLOB.clients)
+				if(!isliving(V.mob))
+					continue
+				var/mob/living/M = V.mob
+				if(M.stat == UNCONSCIOUS || M.stat == DEAD)
+					continue
+				if(!M.latejoin_vore)
+					continue
+				if(!(M.z in using_map.station_levels))
+					continue
+				preds += M
+				pred_names += M.real_name //very cringe
+
+			if(preds.len)
+				var/pred_name = input(C, "Choose a Predator.", "Pred Spawnpoint") as null|anything in pred_names
+				if(!pred_name)
+					return
+				var/index = pred_names.Find(pred_name)
+				var/mob/living/pred = preds[index]
+				vore_spawn_gut = input(C, "Choose a Belly.", "Belly Spawnpoint") as null|anything in pred.vore_organs
+				if(!vore_spawn_gut)
+					return
+				log_admin("[key_name(C)] has requested to vore spawn into [key_name(pred)]")
+				message_admins("[key_name(C)] has requested to vore spawn into [key_name(pred)]")
+
+				var/confirm = alert(pred, "[C.prefs.real_name] is attempting to spawn into your [vore_spawn_gut]. Let them?", "Confirm", "No", "Yes")
+				if(confirm != "Yes")
+					to_chat(C, "<span class='warning'>[pred] has declined your spawn request.</span>")
+					return
+				if(pred.stat == UNCONSCIOUS || pred.stat == DEAD)
+					to_chat(C, "<span class='warning'>[pred] is not conscious.</span>")
+					to_chat(pred, "<span class='warning'>You must be conscious to accept.</span>")
+					return
+				if(!(pred.z in using_map.station_levels))
+					to_chat(C, "<span class='warning'>[pred] is no longer on station.</span>")
+					to_chat(pred, "<span class='warning'>You must be on the station to accept.</span>")
+					return
+				log_admin("[key_name(C)] has vore spawned into [key_name(pred)]")
+				message_admins("[key_name(C)] has vore spawned into [key_name(pred)]")
 			else
-				to_chat(C, "<span class='warning'>Your chosen spawnpoint ([C.prefs.spawnpoint]) is unavailable for the current map. Spawning you at one of the enabled spawn points instead.</span>")
-				spawnpos = null
-		else
+				to_chat(C, "<span class='warning'>No predators were available to accept you.</span>")
+				return
 			spawnpos = spawntypes[C.prefs.spawnpoint]
+		else
+			if(!(C.prefs.spawnpoint in using_map.allowed_spawns))
+				if(fail_deadly)
+					to_chat(C, "<span class='warning'>Your chosen spawnpoint is unavailable for this map and your job requires a specific spawnpoint. Please correct your spawn point choice.</span>")
+					return
+				else
+					to_chat(C, "<span class='warning'>Your chosen spawnpoint ([C.prefs.spawnpoint]) is unavailable for the current map. Spawning you at one of the enabled spawn points instead.</span>")
+					spawnpos = null
+			else
+				spawnpos = spawntypes[C.prefs.spawnpoint]
 
 	//We will return a list key'd by "turf" and "msg"
-	. = list("turf","msg")
+	. = list("turf","msg", "voreny")
+	if(vore_spawn_gut)
+		.["voreny"] = vore_spawn_gut
 	if(spawnpos && istype(spawnpos) && spawnpos.turfs.len)
 		if(spawnpos.check_job_spawning(rank))
 			.["turf"] = spawnpos.get_spawn_position()
